moved update inner bounds
